Show gstreamer errors in wxGStreamerMediaBackend to the user.
If the media file fails to play, we really need to show to the user the
corresponding gstreamer error, otherwise there is no way to understand what
the problem is.

Collect the errors in gst_error_callback() and show them from the main thread
if we get a gstreamer error.