Ensure that names of saved XPMs are valid C identifiers.
The name of the XPM should be based on the file name but can't always be
exactly equal to it as it's possible (and relatively common, e.g.
"foo-24.xpm") to have characters forbidden in identifiers to appear in the
file names. Ensure that we use a valid C identifier here.

Closes #13905.