Don't use delayed destruction for sockets in other threads.
The delayed destruction mechanism is not MT-safe, so using it for wxSocket
objects destroyed from threads other than main resulted in crashes. Luckily,
it is not necessary to use it for such sockets anyhow as they don't risk
receiving any events -- which are only dispatched in the main thread -- and so
can be destroyed immediately.

So do destroy them directly instead of just scheduling for later destruction
when wxSocket::Destroy() is called.