Don't change client size of TLW when centering it in wxGTK.
Setting the frame client size and then calling Centre() on it resulted in the
frame coming up with a wrong client size in wxGTK because the call to Centre()
resulted in calling wxTopLevelWindow::DoSetSize() which unconditionally reset
m_deferShowAllowed to true, undoing setting it to false in DoSetClientSize().

Only reset m_deferShowAllowed if the size was really changed but not if the
window was only moved, as happens when centering it.