Make sure wchar_t CRT functions work on OS X.
In OS X and iOS, wchar_t CRT functions convert to char* and fail under
some locales. The safest fix is to set LC_CTYPE to UTF-8 to ensure that
they can handle any input.

Note that this must be done for any app, Cocoa or console, whether or
not it uses wxLocale.

See http://stackoverflow.com/questions/11713745/why-does-the-printf-family-of-functions-care-about-locale