Use "INNO" env var in Inno Setup script only if it's defined.
Just use the sources directory itself if %INNO% is not defined. In fact, I
think we could always use the source directory, actually, but just in case
there was some reason to do it like this, keep using %INNO% if it is defined.

Also check that WXW_VER environment variable that we use is, in fact, defined.

FIX: Use relative path, not %WXWIN% in ISS script.