Fix light magenta definition in the colour database.
Make it different from magenta itself, otherwise a colour constructed from
"magenta" string was returning "light magenta" as its name.

The exact value of this colour is not well-defined, another commonly used one
seems to be 0xff80ff, but in a lot of places it is used as a synonym for
fuchsia pink which is pretty unambiguously 0xff77ff.

Closes #15628.