Show keyboard focus in generic wxDVC even with single editable column.
Previously, keyboard focus indicator was only shown if there were
multiple editable columns. This was confusing to the user in
applications where the number of editable columns changes: somethings
focus would work, sometimes not. (In reality, it worked in both cases,
but wasn't visually indicated.)

Just always show the indicator if there are any editable columns. Notice
that the indicator is only painted after the user pressed left or right
arrows, thus indicating interest in keyboard focus. Without that, only
the regular indicator is shown (this is the same as before this commit).