Fix harmless C4275 MSVC warning in wxMSW DLL builds.
Since the addition of wxMSWOwnerDrawnButton, DLL builds started generated a
lot of warnings about using this non DLL-exported class as base class for the
DLL-exported wxCheckBox and wxRadioButton.

Simply suppress these warnings as they are harmless in this case because the
base class has no static data, which is the real problem this warning hints at.

Closes #16237.