Provisional patch to support nested event loops in wxQT
Without the patch, the wxExecute (wxAppTraits::RunLoopUntilChildExit) loop.Run() don't process synchonics events (not entering the new loop), so it cannot process exit code and streams (causing a SIGSEGV crash).
The problem is that QCoreApplication::exec() seems to not start a new loop, maybe it should create a real new loop with QEventLoop but then the inner loop don't seems to process events/signals for the outer one (app).
Provisionaly, to get wxExecute working, it calls processEvents() on the main app loop until wxGUIEventLoop::ScheduleExit() is called, terminating the inner loop.
This seems to work for the exec sample, but exec tests keep crashing, will investigate further.