Tweak sizing for general SetWindowVariant() implementation.
Previously, the generic implementation closely mirrored OS X's scaling
factors, with the small size being 3/4 of the normal one and mini 2/3.
This works well with OS X's UI font and font rendering, but less so on
other platforms: wxWINDOW_VARIANT_SMALL is barely readable on Windows
and wxWINDOW_VARIANT_MINI is illegible.  In wxGTK, both sizes are
readable, but don't match small text used in native GNOME interfaces.

Change the sizing to use the same scaling factor of 1.2 between all
variant steps.  This is the same factor that CSS 1.2 defines for
relative sizes and is also used by Pango's (and so GTK+) size scaling in
markup.  This makes wxWINDOW_VARIANT_SMALL identical in size to <small>
markup used in GTK+.

On Windows, this changes default UI font scaling to be less steep and
while wxWINDOW_VARIANT_MINI is still tiny and hard to read (as it is on
OS X), wxWINDOW_VARIANT_SMALL is now readable, thanks to being only 1pt
smaller than the full size (for default GUI font).