Fix error message about ChangeWindowMessageFilter() on startup under XP.
Use wxDynamicLibrary::RawGetSymbol() (from inside wxDL_INIT_FUNC() macro)
instead of GetSymbol() to avoid the error message which was given every time a
new wxFrame was created after wxTaskBarButton changes under XP as
ChangeWindowMessageFilter() function is not available in it.

Also simplify the code a little and don't make pfnChangeWindowMessageFilter
static, we don't need to keep it around.