Fix wxFontDialog exceptions on OS X Yosemite
Opening the font dialog would result in an exception in
RunMixedFontDialog:

-[NSView resetFlags]: unrecognized selector sent to instance 0x6080001285c0

The code in question assumed that NSFontPanel's accessory view either
didn't exist or was created by wx, and casted it to
wxMacFontPanelAccView* without checking. But this assumption is no
longer true on OS X 10.10, the view is apparently pre-set to some
default NSView instance.

Fix the code to check accessoryView's class before treating it as
wxMacFontPanelAccView.