Fix wxStaticText::Disable() to respect text color on OS X
wxStaticText emulates disabled state on OS X by changing text color to
light grey. When re-enabling the control, though, it always set the
color to the standard text color, which broke static texts with a custom
color.

Fix this by keeping track of the original color and restoring it back
when setEnabled:YES is called.