Refactor type checks in wxOSX wxDVC implementation.
Check the type in one place only, before calling MacRender() instead of doing
it in each and every implementation of it.

Also replace wxFAIL_MSG() with wxLogDebug() as the former resulted in a crash
due to assert reentrancy, as usual when asserting inside a wxEVT_PAINT handler
which is constantly called all the time, and so wasn't particularly useful.