Remove files not being compiled from richtext MSBuild project.
Several sources in this project are actually not compiled at all, but are
included from other files, so they don't appear in the master sources list.

Update the project accordingly, for consistency with the earlier versions of
MSVC.

If these files should appear in it, they need to be added to RICHTEXT_HDR
variable.