Allow documents without template nor views be deleted correctly again.
The changes of r78282 didn't take into account the case of a document without
any views nor template and, consequently, without the associated manager. Such
documents were not deleted at all any longer.

Fix this by ensuring that a document is always associated with a manager: if
not with the one specified by its template, then with the global one.